requirements.txt
numpy==2.4.1
python-dotenv==1.2.1
fastapi==0.116.1
openai==1.60.1
uvicorn==0.35.0
